requirements.txt
chainlit
langchain
langchain-community
langchain_openai
openai 
tiktoken
pymupdf 
pinecone-client 
langchain-pinecone
python-dotenv